v2
This script defines a new method called notify_on_arrival in the BusArrivalAPI class. This method takes three arguments: the bus stop code, the bus number, and the desired arrival time. It then enters an infinite loop that makes repeated requests to the LTA DataMall API to get the estimated arrival time for the specified bus at the specified bus stop. If the estimated arrival time matches the desired arrival time, the loop breaks and a notification is printed to the console.